+++ infra/helm-diom/ChangeLog.md
@@ -3,3 +3,4 @@
 ## Unreleased Changes
 * Consistently name operator pod `<release>-operator`
 * Add `operator.logFormat` and `operator.extraEnv` values
+* Add `cluster.spec.storage.ephemeral`