+++ infra/operator/ChangeLog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased Changes
+* Add default crypto provider
+* Add ephemeral storage configuration
 
 ## Version 0.2.0
 * Fix serialization of `OpenTelemetrySpec` fields to use camelCase